requirements.txt
Flask-Login==0.3.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.10.3
Flask-paginate==0.4.5
Pillow==2.7.0
chardet==2.3.0
flask-babelex==0.9.3
flask==0.11.1
phonenumbers==7.5.2
psycopg2==2.6.2
py-bcrypt==0.3
pyasn1==0.1.9
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
uwsgi==2.0.6
simplejson==3.8.2
mandrill==1.0.57
bpython==0.14.1
uwsgitop==0.9
Flask-Cache==0.13.1
gevent==1.1.2
google-api-python-client